🎨 Design: #37 소환사 전적 검색 컴포넌트 디자인
