+++ src/components/Record/Record.module.scss
@@ -8,7 +8,7 @@ $bg-gray: #30333a;
   width: 100%;
 
   .article {
-    width: 1120px;
+    width: 800px;
   }
 }
 
@@ -32,26 +32,16 @@ $bg-gray: #30333a;
   }
 }
 
-@media screen and (max-width: 1120px) {
+@media screen and (max-width: 800px) {
   .main {
-    .article {
-      width: 960px;
-    }
-  }
-}
+    // align-items: flex-start;
 
-@media screen and (max-width: 960px) {
-  .main {
-    .article {
-      width: 720px;
-    }
-  }
-}
-
-@media screen and (max-width: 720px) {
-  .main {
     .article {
       width: 100%;
+    }
+
+    .match {
+      // width: 800px;
     }
   }
 }
