--- src/components/Record/SummonerMatch/MatchInfo.module.scss
@@ -1,75 +1,0 @@
-.card {
-  display: flex;
-  // justify-content: center;
-  flex-direction: column;
-  // align-items: center;
-  height: 100px;
-  border: 1px solid red;
-  margin: 20px;
-  // padding: 20px;
-
-  .title {
-    display: flex;
-    width: 100%;
-    justify-content: space-between;
-
-    .left {
-      .game-result {
-        color: blue;
-      }
-    }
-  }
-
-  .game-stats {
-    display: flex;
-    justify-content: space-between;
-
-    .champion {
-      display: flex;
-
-      .championImg {
-        width: 48px;
-        height: 48px;
-      }
-    }
-    .spell {
-      display: flex;
-      flex-direction: column;
-
-      .spellImg {
-        width: 22px;
-        height: 22px;
-      }
-    }
-
-    .perk {
-      display: flex;
-      flex-direction: column;
-
-      .perkImg {
-        width: 22px;
-        height: 22px;
-      }
-    }
-
-    .items {
-      display: flex;
-      flex-direction: column;
-
-      .list {
-        display: flex;
-
-        .item {
-          width: 22px;
-          height: 22px;
-          background-color: rgb(211 220 233);
-
-          .itemImg {
-            width: 100%;
-            height: 100%;
-          }
-        }
-      }
-    }
-  }
-}
