+++ src/components/Record/SummonerMatch/MatchInfo/MatchInfo.module.scss
@@ -1,0 +1,169 @@
+$bg-color-win: rgb(119, 127, 247);
+$bg-color-loss: rgb(209, 98, 98);
+$text-color: rgb(53, 57, 80);
+
+.card {
+  display: flex;
+  height: 100%;
+  margin: 20px;
+  background-color: rgba($bg-color-loss, 0.3);
+  padding: 10px;
+
+  &.win {
+    background-color: rgba($bg-color-win, 0.3);
+  }
+
+  .title {
+    display: flex;
+    flex-direction: column;
+    // width: 100%;
+    justify-content: space-between;
+    font-size: 12px;
+    margin-right: 20px;
+    white-space: nowrap;
+
+    .game-result {
+      color: rgb(255, 81, 0);
+      font-size: 14px;
+      font-weight: bold;
+    }
+
+    .win {
+      color: rgb(132, 0, 255);
+    }
+  }
+
+  .game-stats {
+    display: flex;
+    width: 100%;
+
+    .left {
+      display: flex;
+      flex-direction: column;
+      margin-right: 20px;
+
+      .champion {
+        display: flex;
+
+        .championImg {
+          width: 56px;
+          height: 56px;
+          border-radius: 10px;
+          margin-right: 5px;
+        }
+      }
+
+      .spell {
+        display: flex;
+        flex-direction: column;
+
+        .spellImg {
+          width: 26px;
+          height: 26px;
+          border-radius: 10px;
+          object-fit: cover;
+          margin-right: 5px;
+          margin-bottom: 2px;
+        }
+      }
+
+      .perk {
+        display: flex;
+        flex-direction: column;
+
+        .perkImg {
+          width: 26px;
+          height: 26px;
+          border-radius: 10px;
+          object-fit: cover;
+          margin-bottom: 2px;
+        }
+      }
+
+      .items {
+        display: flex;
+        margin-top: 10px;
+
+        .list {
+          display: flex;
+
+          .item {
+            width: 26px;
+            height: 26px;
+            margin-right: 5px;
+            background-color: rgb(211 220 233);
+            border-radius: 10px;
+
+            .itemImg {
+              width: 100%;
+              height: 100%;
+              border-radius: 10px;
+            }
+          }
+        }
+      }
+    }
+
+    .right {
+      display: flex;
+      width: 100%;
+      white-space: nowrap;
+
+      .kda {
+        display: flex;
+        flex-direction: column;
+        justify-content: space-around;
+        align-items: center;
+        width: 120px;
+        margin-right: 40px;
+
+        .kdaData {
+          .deaths {
+            color: red;
+          }
+        }
+      }
+
+      .teamMatch {
+        width: 100%;
+      }
+    }
+  }
+}
+
+@media screen and (max-width: 800px) {
+  .card {
+    .title {
+      font-size: 10px;
+    }
+
+    .game-stats {
+      .right {
+        justify-content: center;
+
+        .kda {
+          font-size: 18px;
+          margin-right: 0;
+        }
+
+        .teamMatch {
+          display: none;
+        }
+      }
+    }
+  }
+}
+
+@media screen and (max-width: 450px) {
+  .card {
+    .game-stats {
+      .right {
+        display: none;
+
+        .kda {
+          font-size: 12px;
+        }
+      }
+    }
+  }
+}
