+++ src/components/Record/SummonerMatch/MatchInfo/MatchTeam/MatchTeam.module.scss
@@ -1,0 +1,44 @@
+.teams {
+  display: flex;
+  justify-content: space-around;
+  // width: 300px;
+  width: 100%;
+  height: 100%;
+  font-size: 12px;
+
+  .team {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+
+    .list {
+      display: flex;
+      align-items: center;
+      margin-bottom: 2px;
+      width: 140px;
+
+      .championImg {
+        width: 20px;
+        height: 20px;
+        margin-right: 2px;
+      }
+
+      .link {
+        color: #fff;
+        text-decoration: none;
+        outline: none;
+
+        &:hover,
+        &:active {
+          text-decoration: none;
+          color: #fff;
+          background-color: rgba(aqua, 0.2);
+        }
+      }
+    }
+
+    &:nth-of-type(1) {
+      margin-right: 40px;
+    }
+  }
+}
