+++ src/components/Record/SummonerRank/SummonerRank.module.scss
@@ -92,7 +92,7 @@ $bg-gray: #30333a;
   }
 }
 
-@media screen and (max-width: 600px) {
+@media screen and (max-width: 750px) {
   .section {
     flex-direction: column;
     margin: 20px;
