+++ src/styles/global.scss
@@ -2,6 +2,7 @@
 
 body {
   font-family: 'Roboto', sans-serif;
+  color: white;
 
   @media screen and (max-width: 280px) {
     body {
@@ -29,3 +30,6 @@ body {
   }
 }
 
+main {
+  background-color: #1c1c1f;
+}
